pipeline/requirements.txt
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
playwright>=1.43.0
google-cloud-storage>=2.16.0
python-telegram-bot>=21.0
SQLAlchemy>=2.0.0
